Restructure the signal field validators
