+++ lib/sanbase/signals/types.ex
@@ -2,6 +2,7 @@ defmodule Sanbase.Signal.Type do
   @type trigger_type :: String.t()
   @type channel :: String.t() | list(String.t())
   @type target :: String.t()
+  @type metric :: String.t()
   @type complex_target :: target | list(target) | map()
   @type filtered_target :: %{list: list(), type: trigger_type()}
   @type asset :: String.t()
